Refactoring API of mfg_inspector.py to primarily include a MfgInspectorCallback that can be configured to use a specific converter to convert a test_record to a mfg-inspector compatible protobuf.  The callback can also be configured to save the proto to disk and/or upload the proto to mfg-inspector.
Removed UploadOrOutput class, this functionality was un-used.

PiperOrigin-RevId: 227938986
